+++ src/java/Model/Route.java
@@ -56,7 +56,7 @@ public class Route {
         return ticketPrice;
     }
 
-    public void setTicketPrice(int ticketPrice) {
+    public void setTicketPrice(double ticketPrice) {
         this.ticketPrice = ticketPrice;
     }
